+++ example/pageview.dart
@@ -1,5 +1,4 @@
-import 'package:anlifecycle/anlifecycle.dart';
-import 'package:anlifecycle/src/core/lifecycle_pageview.dart';
+import 'package:an_lifecycle/an_lifecycle.dart';
 import 'package:flutter/material.dart';
 
 import 'lifecycle_printer.dart';