+++ lib/an_lifecycle.dart
@@ -1,4 +1,4 @@
-library anlifecycle;
+library an_lifecycle;
 
 export 'src/core/lifecycle.dart';
 export 'src/core/lifecycle_app.dart';