+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* 变更package为an_lifecycle
+* lifecycle的提供者使用InheritedWidget来保证变更的及时通知
+
 ## 1.0.5
 
 * 修正dispose与!mounted的状态不一致的bug，引起的page的state异常